+++ PrototypeWithAuth/Models/Payment.cs
@@ -12,7 +12,7 @@ namespace PrototypeWithAuth.Models
     {
         [Key]
         public int PaymentID { get; set; }
-        public double Sum { get; set; }
+        public decimal Sum { get; set; }
         
         [DataType(DataType.Date)]
         public DateTime PaymentDate { get; set; }
